refactor(suite-native): dedupe earn fee level thunk copies
